# FreeTimeApp/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "tiempo-libre",
    "version": "1.0"
  },
  "paths": {
...
    "/api/notificaciones/obtener": {
      "post": {
        "tags": [
          "Notificaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObtenerNotificacionesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ObtenerNotificacionesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ObtenerNotificacionesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notificaciones/marcar-leida/{notificacionId}": {
      "patch": {
        "tags": [
          "Notificaciones"
        ],
        "parameters": [
          {
            "name": "notificacionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notificaciones/archivar/{notificacionId}": {
      "patch": {
        "tags": [
          "Notificaciones"
        ],
        "parameters": [
          {
            "name": "notificacionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EstatusNotificacionEnum": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ObtenerNotificacionesRequest": {
        "type": "object",
        "properties": {
          "tipoNotificacion": {
            "$ref": "#/components/schemas/TiposDeNotificacionEnum"
          },
          "estatus": {
            "$ref": "#/components/schemas/EstatusNotificacionEnum"
          },
          "areaId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "grupoId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fechaInicio": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaFin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pagina": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "tamañoPagina": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "ordenarPor": {
            "type": "string",
            "nullable": true
          },
          "direccionOrden": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TiposDeNotificacionEnum": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}